chore: restructuring directories
