+++ components/features/FilterMenu/FilterMenu.tsx
@@ -1,8 +1,8 @@
-import ISectionObj from "../../interfaces/ISectionObj";
-import ICategoryObj from "../../interfaces/ICategoryObj";
+import ISectionObj from "../../../interfaces/ISectionObj";
+import ICategoryObj from "../../../interfaces/ICategoryObj";
 
-import handleFilterByCategory from "../../util/handleFilterByCategory";
-import handleFilterBySection from "../../util/handleFilterBySection";
+import handleFilterByCategory from "../../../util/handleFilterByCategory";
+import handleFilterBySection from "../../../util/handleFilterBySection";
 
 const FilterMenu: React.FC<{
   sections?: ISectionObj[];
