+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
-import Search from "../components/search/Search/Search";
+import Search from "../components/features/Search/Search";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
