+++ pages/repo/[[...slug]].tsx
@@ -4,13 +4,13 @@ import type {
   GetStaticPropsContext,
   NextPage,
 } from "next";
-import RepoPage from "../../components/RepoPage/RepoPage";
-import SectionPage from "../../components/SectionPage/SectionPage";
-import CategoryPage from "../../components/CategoryPage/CategoryPage";
-import PostPage from "../../components/PostPage/PostPage";
+import RepoPage from "../../components/pages/RepoPage/RepoPage";
+import SectionPage from "../../components/pages/SectionPage/SectionPage";
+import CategoryPage from "../../components/pages/CategoryPage/CategoryPage";
+import PostPage from "../../components/pages/PostPage/PostPage";
 import IPost from "../../interfaces/IPost";
-import { getContentByDynamicPage } from "../../util/pages/getContentByDynamicPage";
-import { getDynamicPagePaths } from "../../util/pages/getDynamicPagePaths";
+import { getContentByDynamicPage } from "../../util/getContentByDynamicPage";
+import { getDynamicPagePaths } from "../../util/getDynamicPagePaths";
 
 const DynamicPage: NextPage<{ slug: string; content?: string }> = ({
   slug,
